requirements.txt
discord.py==2.3.2
datetime==5.4
google-api-python-client==1.7.2
google-auth==2.14.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
isodate==0.6.1
requests==2.31.0
